Isolate useUpgradeCheck tests for Bun coverage
